cleaning main.py
